[FEAT] restructured codebase, added linting in this package via its own ruleset +typescript in js mode, removed spellchecker plugin as it's bad, added comments plugin, removed dependency on airbnb as we'll define our own rules
