+++ .eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+    extends: [
+        './packages/eslint-config-assignar-base/index.js'
+    ]
+}
