--- flavours/node/index.js
@@ -1,11 +1,0 @@
-module.exports = {
-    extends: [
-        'eslint-config-assignar-base',
-    ].map(require.resolve),
-    env: {
-        node: true,
-    },
-    settings: {
-        'import/resolver': 'babel-module',
-    },
-}
