--- flavours/react/package.json
@@ -1,5 +1,0 @@
-{
-    "dependencies": {
-        "eslint-config-assignar": ""
-    }
-}
