--- flavours/react/plugin/import.js
@@ -1,5 +1,0 @@
-module.exports = {
-    // No Node.js builtin modules
-    // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-nodejs-modules.md
-    'import/no-nodejs-modules': 'error',
-}
