+++ packages/eslint-config-assignar-base/.npmignore
@@ -1,0 +1,5 @@
+**/*
+!./plugins/**/*
+!./index.js
+!./package.json
+!./README.md
