+++ packages/eslint-config-assignar-base/README.md
@@ -1,0 +1,6 @@
+# eslint-config-assignar-base
+
+This is the base eslint configuration for @assignar.
+
+This package contains common configuration for the main packages.
+Generally you shouldn't install this package directly, as it will be installed as a dependency of the main packages.
