+++ packages/eslint-config-assignar-base/package.json
@@ -1,24 +1,20 @@
 {
-    "name": "eslint-config-assignar",
-    "private": true,
+    "name": "eslint-config-assignar-base",
     "version": "0.0.0",
-    "description": "assignar's eslint config",
+    "license": "MIT",
     "repository": "https://github.com/assignar/eslint-config-assignar.git",
     "author": "Brad Zacher <brad.zacher@gmail.com>",
-    "license": "MIT",
-    "devDependencies": {
-        "@types/node": "^10.3.4",
-        "eslint": "^4.19.1",
+    "main": "index.js",
+    "peerDependencies": {
+        "eslint": "^4.19.0"
+    },
+    "dependencies": {
         "eslint-plugin-dependencies": "^2.4.0",
         "eslint-plugin-eslint-comments": "^2.0.2",
         "eslint-plugin-import": "^2.9.0",
         "eslint-plugin-jest": "^21.15.0",
         "eslint-plugin-typescript": "^0.12.0",
         "eslint-restricted-globals": "^0.1.1",
-        "typescript": "^2.9.2",
         "typescript-eslint-parser": "^16.0.0"
-    },
-    "scripts": {
-        "lint": "tsc -p ./jsconfig.json && eslint --cache ./packages"
     }
 }
