+++ packages/eslint-config-assignar-base/plugins/import.js
@@ -1,0 +1,204 @@
+module.exports = {
+    rules: {
+        // ensure default import coupled with default export
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/default.md
+        'import/default': 'error',
+
+        // disallow invalid exports, e.g. multiple defaults
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/export.md
+        'import/export': 'error',
+
+        // This rule enforces that all exports are declared at the bottom of the file.
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/exports-last.md
+        'import/exports-last': 'off',
+
+        // Ensure consistent use of file extension within the import path
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/extensions.md
+        'import/extensions': ['error', 'ignorePackages', {
+            js: 'never',
+            jsx: 'never',
+            ts: 'never',
+            tsx: 'never',
+        }],
+
+        // disallow non-import statements appearing before import statements
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/first.md
+        'import/first': 'error',
+
+        // Reports when named exports are not grouped together in a single export declaration
+        // or when multiple assignments to CommonJS module.exports or exports object are present
+        // in a single file.
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/group-exports.md
+        'import/group-exports': 'off',
+
+        // disallow non-import statements appearing before import statements
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/imports-first.md
+        // deprecated: use `import/first`
+        'import/imports-first': 'off',
+
+        // Forbid modules to have too many dependencies
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/max-dependencies.md
+        'import/max-dependencies': ['off', { max: 10 }],
+
+        // ensure named imports coupled with named exports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/named.md#when-not-to-use-it
+        // covered by typescript
+        'import/named': 'off',
+
+        // enforces names exist at the time they are dereferenced
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/namespace.md
+        'import/namespace': 'off',
+
+        // Require a newline after the last import/require in a group
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/newline-after-import.md
+        'import/newline-after-import': 'error',
+
+        // Forbid import of modules using absolute paths
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-absolute-path.md
+        'import/no-absolute-path': 'error',
+
+        // disallow AMD require/define
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-amd.md
+        'import/no-amd': 'error',
+
+        // Reports if a module's default export is unnamed
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-anonymous-default-export.md
+        'import/no-anonymous-default-export': ['off', {
+        allowArray: false,
+        allowArrowFunction: false,
+        allowAnonymousClass: false,
+        allowAnonymousFunction: false,
+        allowLiteral: false,
+        allowObject: false,
+        }],
+
+        // disallow require()
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-commonjs.md
+        'import/no-commonjs': 'off',
+
+        // forbid default exports. this is a terrible rule, do not use it.
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-default-export.md
+        'import/no-default-export': 'off',
+
+        // disallow use of jsdoc-marked-deprecated imports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-deprecated.md
+        'import/no-deprecated': 'off',
+
+        // disallow duplicate imports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-duplicates.md
+        'import/no-duplicates': 'error',
+
+        // Forbid require() calls with expressions
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-dynamic-require.md
+        'import/no-dynamic-require': 'error',
+
+        // Forbid the use of extraneous packages
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-extraneous-dependencies.md
+        'import/no-extraneous-dependencies': ['error', {
+            devDependencies: true,
+            peerDependencies: true,
+            optionalDependencies: false,
+        }],
+
+        // prevent importing the submodules of other modules
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-internal-modules.md
+        'import/no-internal-modules': ['off', {
+            allow: [],
+        }],
+
+        // Forbid mutable exports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-mutable-exports.md
+        'import/no-mutable-exports': 'error',
+
+        // do not allow a default import name to match a named export
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-named-as-default.md
+        'import/no-named-as-default': 'error',
+
+        // warn on accessing default export property names that are also named exports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-named-as-default-member.md
+        'import/no-named-as-default-member': 'error',
+
+        // Prevent importing the default as if it were named
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-named-default.md
+        'import/no-named-default': 'error',
+
+        // disallow namespace imports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-namespace.md
+        'import/no-namespace': 'off',
+
+        // No Node.js builtin modules
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-nodejs-modules.md
+        'import/no-nodejs-modules': 'off',
+
+        // Restrict which files can be imported in a given folder
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-restricted-paths.md
+        'import/no-restricted-paths': 'off',
+
+        // Forbid a module from importing itself
+        // https://github.com/benmosher/eslint-plugin-import/blob/44a038c06487964394b1e15b64f3bd34e5d40cde/docs/rules/no-self-import.md
+        'import/no-self-import': 'error',
+
+        // Prevent unassigned imports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-unassigned-import.md
+        // importing for side effects is perfectly acceptable, if you need side effects.
+        'import/no-unassigned-import': 'off',
+
+        // ensure imports point to files/modules that can be resolved
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-unresolved.md
+        'import/no-unresolved': ['error', {
+            commonjs: true,
+            caseSensitive: true,
+        }],
+
+        // Ensures that there are no useless path segments
+        // https://github.com/benmosher/eslint-plugin-import/issues/1032
+        'import/no-useless-path-segments': 'error',
+
+        // Forbid Webpack loader syntax in imports
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-webpack-loader-syntax.md
+        'import/no-webpack-loader-syntax': 'error',
+
+        // Enforce a convention in module import order
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/order.md
+        'import/order': ['error', {
+            groups: [
+                'builtin',
+                'external',
+                'internal',
+                'parent',
+                'sibling',
+                'index',
+            ],
+            'newlines-between': 'always',
+        }],
+
+        // Require modules with a single export to use a default export
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/prefer-default-export.md
+        'import/prefer-default-export': 'error',
+
+        // Warn if a module could be mistakenly parsed as a script by a consumer
+        // leveraging Unambiguous JavaScript Grammar
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/unambiguous.md
+        // this should not be enabled until this proposal has at least been *presented* to TC39.
+        // At the moment, it's not a thing.
+        'import/unambiguous': 'off',
+    },
+    settings: {
+        'import/extensions': [
+            'js',
+            'jsx',
+            'ts',
+            'tsx',
+        ],
+        'import/parsers': {
+            'typescript-eslint-parser': [
+                'ts',
+                'tsx',
+            ],
+        },
+        'import/ignore': [
+            'node_modules',
+            '\\.(coffee|scss|css|less|hbs|svg|json)$',
+        ],
+    },
+}
