+++ packages/eslint-config-assignar-node/.npmignore
@@ -1,0 +1,5 @@
+**/*
+!./plugins/**/*
+!./index.js
+!./package.json
+!./README.md
