+++ packages/eslint-config-assignar-node/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+    extends: [
+        require.resolve('eslint-config-assignar-base'),
+    ],
+    env: {
+        node: true,
+    },
+}
