+++ packages/eslint-config-assignar-react/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "eslint-config-assignar-react",
+    "version": "0.0.0",
+    "license": "MIT",
+    "repository": "https://github.com/assignar/eslint-config-assignar.git",
+    "author": "Brad Zacher <brad.zacher@gmail.com>",
+    "main": "index.js",
+    "peerDependencies": {
+        "eslint": "^4.19.0"
+    },
+    "dependencies": {}
+}
