+++ packages/eslint-config-assignar-react/plugins/import.js
@@ -1,0 +1,7 @@
+module.exports = {
+    rules: {
+        // No Node.js builtin modules
+        // https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-nodejs-modules.md
+        'import/no-nodejs-modules': 'error',
+    },
+}
