+++ types/eslint-restricted-globals.d.ts
@@ -1,0 +1,4 @@
+declare module 'eslint-restricted-globals' {
+    const globals : string[]
+    export = globals
+}
